Move type hint tests to their own module
